requirements.txt
anthropic==0.49.0
httpx==0.28.1
beautifulsoup4==4.13.3
lxml==5.3.0
pyyaml==6.0.2
rich==14.0.0
